Modify raw data
The first row of the csv files are supposed to be used as column names, but spaces are detected so that Stata treats them as label for column names rather than column names. I modified the first row for the three csv files. Pure number also could not serve as column names, so I added 'year' before the year numbers using code.